hwpacks/linaro-arndale: add arndale-ethernet package to set the MAC address (LP: #1097265)
when the kernel boots and asix driver is loaded,
the driver tries to read the mac address but fails:
asix 1-3.2.4:1.0 (unregistered net_device): invalid hw address, using random

the driver fallback is to assign a random mac address.
it's causing dhcpd to run out of leases.

The package include an udev rule to always set the MAC address to
00:40:5c:26:0a:5b.

Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>